rename lib to app
