+++ app/models/charity.rb
