+++ app/models/product.rb
